orchestrator_service/requirements.txt
# Orchestrator Service Dependencies

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP client
aiohttp==3.9.1

# Data validation
pydantic==2.5.0

# Environment
python-dotenv==1.0.0

# Utilities
aiofiles==23.2.1
psutil==5.9.6
